pyproject.toml
[tool.poetry]
name = "Voicecord"
version = "1.0.0"
description = "A code that can make your discord account 24/7 on voice channels!"
authors = ["SealedSaucer"]

[tool.poetry.dependencies]
python = ">=3.10.0,<3.11"
numpy = "^1.22.2"
replit = "^3.2.4"
Flask = "^2.2.0"
urllib3 = "^1.26.12"
requests = "2.31.0"
websocket-client = "^1.9.0"

[tool.poetry.dev-dependencies]
debugpy = "^1.6.2"
replit-python-lsp-server = {extras = ["yapf", "rope", "pyflakes"], version = "^1.5.9"}

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
